Removed changes in variant.h completely and reduced changes needed in variant.cpp.
The USART0_Handler function in variant.cpp has to be commented out for this to work, I haven't found a way to change the handler for that IRQ yet.

Also, Implemented Stream with the DigiFi class, you can now do wifi.write(), wifi.available(), wifi.read() etc just like with the Serial classes.

Signed-off-by: ags131 <ashumann13@gmail.com>